Orthodox Canonical Form for ex02

+++ ex02/include/PmergeMe.hpp
@@ -47,7 +47,7 @@ public:
 
 // Utility functions
 void print(std::vector<int> &, std::deque<int> &);
-void caluclateTime(std::vector<int> &, std::deque <int> &, int &, int &);
+void caluclateTime(std::vector<int> &, std::deque <int> &, double &, double &);
 int midPoint(int start, int end);
 
 #endif
